add cfg and dis analysis files
