+++ core/src/analysis/mod.rs
@@ -1,1 +1,3 @@
+pub mod cfg;
+pub mod dis;
 pub mod pe;
